a. utc and decimal precision bug / b. dont use width in UI pages

+++ Code/BackEnd/Libs/Infrastructure/Models/DTOs/ConsumerReadingData.cs
@@ -1,7 +1,7 @@
 namespace Simulators.Models;
 
 public record ConsumerReadingData(
-    int Power,
+    decimal Power,
     DateTime Timestamp, //Utc
     int AddressId
 );
